Implement inventory-centric tracking, credit ledger, salt-based substitutions, and receipt updates

+++ lib/core/services/shortbook_service.dart
@@ -28,7 +28,7 @@ class ShortbookService {
         result.add(ShortbookItem(
           product: product,
           currentStock: stock,
-          threshold: product.minStockThreshold,
+          threshold: product.minStockThreshold.toDouble(),
         ));
       }
     }
